Split virtio-fs transport and protocol ownership
